pyproject.toml
[tool.poetry]
name = "niver65"
version = "0.1.0"
description = "For sidon 65's birthday"
authors = ["Sidon Duarte <sidoncd@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12.2"
python-json-logger = "2.0.7"
requests = "2.31.0"
fastapi = "0.105.0"
uvicorn = "0.24.0"
tenacity = "8.2.3"
behave = "1.2.6"
nose = "1.3.7"
httpx = "^0.25.2"
sqlalchemy = "^2.0.25"
psycopg2 = "^2.9.9"
asyncpg = "^0.29.0"
pyyaml = "^6.0.1"
pydantic = {extras = ["email"], version = "^2.7.1"}

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-cov = "^4.1.0"
pandas = "^2.1.4"
odfpy = "^1.4.1"
faker = "^22.5.1"
factory-boy = "^3.3.0"
ipython = "^8.20.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"